Completed upto Shipment log list

+++ modules/Common/DTO/OtherOp/cargo_otherop_dto.cs
@@ -84,10 +84,10 @@ namespace Common.DTO.OtherOp
         public string? oth_it_date { get; set; }= "";
         public string? oth_it_port { get; set; } = "";
         public decimal? oth_teu { get; set; } = 0;
-        public decimal? oth_20 { get; set; } = 0;
-        public decimal? oth_40 { get; set; } = 0;
-        public decimal? oth_40hq { get; set; } = 0;
-        public decimal? oth_45 { get; set; } = 0;
+        public int? oth_20 { get; set; } = 0;
+        public int? oth_40 { get; set; } = 0;
+        public int? oth_40hq { get; set; } = 0;
+        public int? oth_45 { get; set; } = 0;
         public int? oth_container_tot { get; set; } = 0;
         public int? rec_files_count { get; set; }
         public string? rec_files_attached { get; set; }
